chore: rename example app
