+++ example/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">RNMaps Example</string>
+    <string name="app_name">rnmshowcase</string>
 </resources>
