+++ example/ios/rnmshowcase.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
