+++ example/ios/rnmshowcase/Images.xcassets/AppIcon.appiconset/Contents.json
