+++ example/ios/rnmshowcase/Images.xcassets/Contents.json
