+++ example/ios/rnmshowcaseTests/Info.plist
