+++ example/ios/rnmshowcaseTests/rnmshowcaseTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface exampleTests : XCTestCase
+@interface rnmshowcaseTests : XCTestCase
 
 @end
 
-@implementation exampleTests
+@implementation rnmshowcaseTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL (^)(UIView *view))test
 {
