+++ example/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "example",
+  "name": "rnmshowcase",
   "version": "0.0.1",
   "private": true,
   "scripts": {
